Add files via upload
Game for checking a user memory:  a user should start the game by clicking on any button, then the next button which should be pressed is shown in lighted up mode, if  a user select this chosen button he /she will level up. for following level he/she should remember the pattern of the lighted buttons and the game becomes more complicated as it level up due to the complex pattern.